refactor: :art: improve the architecture with clean architecture
